move proto utilities to separate file
