+++ trellis/core/publisher.hpp
@@ -20,6 +20,7 @@
 
 #include <ecal/msg/protobuf/publisher.h>
 
+#include "trellis/core/proto_utils.hpp"
 #include "trellis/core/time.hpp"
 #include "trellis/core/timestamped_message.pb.h"
 
@@ -68,54 +69,25 @@ class DynamicPublisherImpl : public eCAL::CMsgPublisher<google::protobuf::Messag
    * @param msg Protobuf message
    **/
   DynamicPublisherImpl(const std::string& topic_name, std::shared_ptr<google::protobuf::Message> msg)
-      : CMsgPublisher<google::protobuf::Message>(topic_name, GetTypeName(msg.get()), GetDescription(msg.get())),
+      : CMsgPublisher<google::protobuf::Message>(topic_name, proto_utils::GetTypeName(msg.get()),
+                                                 proto_utils::GetDescription(msg.get())),
         msg_{msg_} {}
 
   DynamicPublisherImpl(const std::string& topic_name, const std::string& proto_type_name)
-      : CMsgPublisher<google::protobuf::Message>(topic_name, GetTypeName(CreateMessageByName(proto_type_name).get()),
-                                                 GetDescription(CreateMessageByName(proto_type_name).get())),
-        msg_{CreateMessageByName(proto_type_name)} {}
+      : CMsgPublisher<google::protobuf::Message>(
+            topic_name, proto_utils::GetTypeName(proto_utils::CreateMessageByName(proto_type_name).get()),
+            proto_utils::GetDescription(proto_utils::CreateMessageByName(proto_type_name).get())),
+        msg_{proto_utils::CreateMessageByName(proto_type_name)} {}
 
   DynamicPublisherImpl(const DynamicPublisherImpl&) = delete;
   DynamicPublisherImpl& operator=(const DynamicPublisherImpl&) = delete;
   DynamicPublisherImpl(DynamicPublisherImpl&&) = default;
   DynamicPublisherImpl& operator=(DynamicPublisherImpl&&) = default;
 
-  static std::string GetTypeName(const google::protobuf::Message* msg) { return ("proto:" + msg->GetTypeName()); }
-  std::string GetTypeName() const override { return GetTypeName(msg_.get()); }
-
-  static std::shared_ptr<google::protobuf::Message> CreateMessageByName(const std::string& type_name) {
-    const google::protobuf::Descriptor* desc =
-        google::protobuf::DescriptorPool::generated_pool()->FindMessageTypeByName(type_name);
-
-    if (desc == nullptr) {
-      std::stringstream errmsg;
-      errmsg << "Unable to find message type " << type_name << std::endl;
-      throw std::runtime_error(errmsg.str());
-    }
-    std::shared_ptr<google::protobuf::Message> message{
-        google::protobuf::MessageFactory::generated_factory()->GetPrototype(desc)->New()};
-
-    if (message == nullptr) {
-      std::stringstream errmsg;
-      errmsg << "Unable to create new message from type " << type_name << std::endl;
-      throw std::runtime_error(errmsg.str());
-    }
-    return message;
-  }
+  std::string GetTypeName() const override { return proto_utils::GetTypeName(msg_.get()); }
 
  private:
-  static std::string GetDescription(const google::protobuf::Message* msg) {
-    const google::protobuf::Descriptor* desc = msg->GetDescriptor();
-    google::protobuf::FileDescriptorSet pset;
-    if (eCAL::protobuf::GetFileDescriptor(desc, pset)) {
-      std::string desc_s = pset.SerializeAsString();
-      return (desc_s);
-    }
-    return ("");
-  }
-
-  std::string GetDescription() const override { return GetDescription(msg_.get()); }
+  std::string GetDescription() const override { return proto_utils::GetDescription(msg_.get()); }
   size_t GetSize(const google::protobuf::Message& msg) const { return ((size_t)msg.ByteSizeLong()); }
   bool Serialize(const google::protobuf::Message& msg, char* buffer_, size_t size_) const {
     return (msg.SerializeToArray((void*)buffer_, (int)size_));
